Refactor data module & add spacy ml module
