+++ .pylintrc
@@ -6,6 +6,7 @@
 extension-pkg-allow-list=
       pydantic,
       celery,
+      spacy,
       lxml
 
 # A comma-separated list of package or module names from where C extensions may
