+++ doc/sources/api/data.rst
@@ -5,3 +5,27 @@ eyes data module
 
 .. automodule:: eyes.data
     :members:
+
+eyes.data.ptt
+-------------
+
+.. automodule:: eyes.data.ptt
+    :members:
+
+eyes.data.dcard
+---------------
+
+.. automodule:: eyes.data.dcard
+    :members:
+
+eyes.data.stats
+---------------
+
+.. automodule:: eyes.data.stats
+    :members:
+
+eyes.data.spacy
+---------------
+
+.. automodule:: eyes.data.spacy
+    :members:
