+++ pyproject.toml
@@ -23,6 +23,8 @@ uvicorn = "^0.15.0"
 graphene = "^2.1.9"
 fastapi = "^0.68.1"
 graphene-sqlalchemy = "^2.3.0"
+spacy = "^3.1.2"
+skweak = "^0.2.13"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.0"
