+++ tests/integration/test_celery.py
@@ -4,7 +4,7 @@ from typing import Dict, List
 
 import pytest
 import sqlalchemy as sa
-from sqlalchemy.orm import Session, sessionmaker
+from sqlalchemy.orm import Session, scoped_session, sessionmaker
 
 from eyes.celery.crawler.tasks import (
     crawl_dcard_board_list,
@@ -34,24 +34,14 @@ class TestCrawler:
         ]
 
     @pytest.fixture
-    def session(
-        self,
-        tables,
-    ):
+    def session(self):
         db_config = MySQLConfig()
         engine = sa.create_engine(
             f'mysql://{db_config.user}:{db_config.password}@{db_config.host}:{db_config.port}/{db_config.database}?charset=utf8mb4'
         )
-        session = sessionmaker(engine)()
-        session.execute('SET FOREIGN_KEY_CHECKS=0')
-        session.commit()
-        for tbl in tables:
-            session.execute(f'TRUNCATE TABLE {tbl}')
-            session.commit()
+        session_factory = sessionmaker(engine)
+        session = scoped_session(session_factory)()
         yield session
-        for tbl in tables:
-            session.execute(f'TRUNCATE TABLE {tbl}')
-            session.commit()
         session.close()
 
     def test_ptt_celery_crawler(
